Initial setup HTML and CSS
It includes the initial setup of HTML and CSS files for my personal portfolio website project. The HTML file contains the basic structure of the website, including sections for the header, main content, and footer. The CSS file defines the global styles, such as font styles, colors, and layout settings, to establish the visual appearance of the website. These files lay the foundation for further development and customization of the portfolio website